# docs/swagger.yaml
basePath: /api/v1
definitions:
  response.ErrorBody:
    properties:
      error:
        type: integer
      error_description:
        type: string
    type: object
  user.CreateRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      role:
        enum:
        - user
        - admin
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  user.UpdateRequest:
    properties:
      full_name:
        type: string
      is_active:
        type: boolean
      role:
        enum:
        - user
        - admin
        type: string
    type: object
info:
  contact: {}
  description: API server for go-api-template
  title: go-api-template API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users with optional search and role filter
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search term
        in: query
        name: q
        type: string
      - description: Filter by role
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with hashed password
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: Soft-delete a user by ID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a single user by their UUID
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user's details
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"